Split all public types into public headers
Split all from single globally included header file, into set of public header files.
